Changed the flows that deploy the SharePoint lists to include the updated columns. Updated Building access.zip with latest changes for the issues described.
Issue 1: When on one day the maximum capacity is booked, you can't book on another day where there is already a booking. When there is no booking on that day you can still book.
Issue 2: When you withdraw a request, the other records created for the same request (depends on the timeslots you have) they stay active, so the app sees this day as still booked. It needs to inactivate all records for the same request.

Fixes Done
- Changed the flows that deploy the SharePoint lists to include the updated columns
- Updated Building access.zip with latest changes for the issues described above